+++ changelog.md
@@ -7,6 +7,7 @@
 * Update the CSIRO renderer to v0.40.0
 * Enhance the POC $extract Questionnaire tester based on initial testing
 * in the FML test page, show the output tab (if not already visible) after running the test
+* questionnaire test page improved outcome issue handling for save/extract
 
 ## 12 September 2024 (dev)
 * Update HealthSamurai Aidbox closure fhirpath engine description